Conflicts:
	app/models/search_criterion.rb

+++ lib/open_chain/custom_handler/fenix_invoice_parser.rb
@@ -10,13 +10,25 @@ module OpenChain
           next if row.length==0
           my_invoice_number = row[3]
           if last_invoice_number!=my_invoice_number && !rows.empty?
-            self.new rows, opts
+            process_invoice_rows rows, opts
             rows = []
           end
           rows << row
           last_invoice_number = my_invoice_number
         end
-        self.new rows, opts unless rows.empty?
+        process_invoice_rows rows, opts unless rows.empty?
+      end
+
+      def self.process_invoice_rows rows, opts
+        begin
+          self.new rows, opts
+        rescue
+          $!.log_me ["Failed to process Fenix Invoice # #{get_invoice_number(rows.first)}" + (opts[:key] ? " from file '#{opts[:key]}'" : "") + "."]
+        end
+      end
+
+      def self.get_invoice_number row
+        row[3].strip
       end
       
       #don't call this, use the static parse method
@@ -30,30 +42,43 @@ module OpenChain
             line = add_detail(invoice, r)
             invoice.invoice_total += line.charge_amount unless line.charge_type=='D'
           end
-          ent = Entry.find_by_source_system_and_broker_reference invoice.source_system, invoice.broker_reference
+
+          ent = Entry.find_by_source_system_and_broker_reference(invoice.source_system, invoice.broker_reference)
           invoice.entry = ent if ent
           invoice.save!
         end
       end
 
       private 
-      def make_header row
-        inv = BrokerInvoice.where(:source_system=>'Fenix',:invoice_number=>row[3].strip).first_or_create!
-        inv.broker_invoice_lines.destroy_all #clear existing lines
-        inv.broker_reference = row[9].strip
-        inv.currency = row[10].strip
-        inv.invoice_date = Date.strptime row[0].strip, '%m/%d/%Y'
-        inv
-      end
+        def make_header row
+          broker_reference = safe_strip row[9]
 
-      def add_detail invoice, row
-        charge_code = row[6].strip
-        charge_code_prefix = charge_code.split(" ").first
-        charge_code = charge_code_prefix if charge_code_prefix.match /^[0-9]*$/
-        line = invoice.broker_invoice_lines.build(:charge_description=>row[7].strip,:charge_code=>charge_code,:charge_amount=>BigDecimal(row[8].strip))
-        line.charge_type = (['20','21'].include?(line.charge_code) ? 'D' : 'R')
-        line
-      end
+          # We need a broker reference in the system to link to an entry, so that we can then know which 
+          # customer the invoice belongs to
+          if broker_reference.blank?
+            raise "Invoice # #{FenixInvoiceParser.get_invoice_number(row)} is missing a broker reference number."
+          end
+
+          inv = BrokerInvoice.where(:source_system=>'Fenix',:invoice_number=>FenixInvoiceParser.get_invoice_number(row)).first_or_create!
+          inv.broker_invoice_lines.destroy_all #clear existing lines
+          inv.broker_reference = broker_reference
+          inv.currency = safe_strip row[10]
+          inv.invoice_date = Date.strptime safe_strip(row[0]), '%m/%d/%Y'
+          inv
+        end
+
+        def add_detail invoice, row
+          charge_code = safe_strip row[6]
+          charge_code_prefix = charge_code.split(" ").first
+          charge_code = charge_code_prefix if charge_code_prefix.match /^[0-9]*$/
+          line = invoice.broker_invoice_lines.build(:charge_description=>safe_strip(row[7]),:charge_code=>charge_code,:charge_amount=>BigDecimal(safe_strip(row[8])))
+          line.charge_type = (['20','21'].include?(line.charge_code) ? 'D' : 'R')
+          line
+        end
+
+        def safe_strip val
+          return val.blank? ? val : val.strip
+        end
     end
   end
 end
